Import np2 r2847 trunk.

+++ common/lstarray.c
@@ -114,7 +114,7 @@ UINT listarray_getpos(LISTARRAY laHandle, void *vpItem) {
 	while(laHandle) {
 		UINT8 *p = (UINT8 *)(laHandle + 1);
 		for (i=0; i<laHandle->items; i++) {
-			if ((long)p == (long)vpItem) {
+			if (p == vpItem) {
 				return(pos + i);
 			}
 			p += laHandle->listsize;
